ozon-referral-bot/requirements.txt
python-telegram-bot==20.7
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.7
redis==5.0.1
requests==2.31.0
python-dotenv==1.0.0
apscheduler==3.10.4
loguru==0.7.2
pydantic==2.5.2